src/third_party/blink/web_tests/display-lock/activation/focus-empty-layout.html

<!doctype HTML>

<!--
Focus on a div that doesn't have style/layout value yet.
-->

<div id="container" style ="contain:style layout">
</div>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<script>
promise_test((t) => {
  const acquirePromise = container.displayLock.acquire({ timeout: Infinity, activatable: true });
  return new Promise((resolve, reject) => {
    const focusable = document.createElement("div");
    focusable.tabIndex = 0;

    const eventPromise = new Promise((resolve, reject) => {
      container.onbeforeactivate = (e) => {
        t.step(() => assert_equals(e.target, container));
        t.step(() => assert_equals(e.activatedElement, focusable));
        resolve();
      };
    });

    acquirePromise.then(() => {
      container.appendChild(focusable);
      focusable.focus();
      eventPromise.then(() => {
        t.step(() => assert_equals(document.activeElement, focusable));
        resolve();
      });
    });
  });
}, "Activating locked element through tabindex navigation fires beforeactivate, focuses element");
</script>
